requirements.consumer.txt
kafka-python==2.0.2
numpy==1.23.5
pandas==1.5.3
